Remove support for legacy python versions (#69)
* Remove legacy asyncio.ensure_future support from AsyncParameterTree

* Refactor API handlers to remove python 2.7 support

This commit refactors the API handlers to remove python 2.7 support by no
longer requiring separate sync/async implementations. The underlying BaseApiHandler
class is also merged into a single ApiHandler that provides async verb methods
compatible with both sync and async adapters. Test cases are also refactored and
simplified to match.

* Refactor routes/api.py to separate adapter list and API version handlers

* Remove redundant python 2.7 code from utils.py and refactor

* Rename remaining py3 test modules

* Remove singleton behaviour from system status and info adapters

Allows the legacy future package to be dropped from dependencies.

* Remove version conditonal import from config/parser.py

* Remove remaining python version checks from test suite

* Fix remaining missing test coverage on BaseParameterTree

Test branch replace fails on immutable tree (and remove excessive
caps from exception. Tsk, tsk Alan ;-) )

* Extend support and test envs to include python 3.13
