+++ pyproject.toml
@@ -12,6 +12,7 @@ classifiers = [
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
 ]
 description = "ODIN detector control system"
 dynamic = ["version"]
@@ -22,9 +23,7 @@ authors = [
 ]
 
 dependencies = [
-    "tornado>=4.3",
-    "future",
-    "pyzmq>=17.1.0",
+    "tornado>=6.0",
     "psutil>=5.0"
 ]
 
@@ -32,8 +31,10 @@ dependencies = [
 dev = [
     "requests",
     "tox",
-    "pytest-asyncio<0.23",
+    "pytest",
+    "pytest-asyncio",
     "pytest-cov",
+    "ruff"
 ]
 sync_proxy = [
     "requests"
@@ -63,7 +64,7 @@ legacy_tox_ini = """
 # tox test configuration for odin-control
 
 [tox]
-envlist = clean,py{38,39,310,311,312},py{38}-pygelf,report
+envlist = clean,py{38,39,310,311,312,313},py{38}-pygelf,report
 
 [gh-actions]
 python =
@@ -72,21 +73,23 @@ python =
     3.10: py310
     3.11: py311
     3.12: py312
+    3.13: py313
 
 [testenv]
 deps =
     pytest
     pytest-cov
-    pytest-asyncio<0.23
+    py{38,38-pygelf}: pytest-asyncio<0.23
+    py{39,310,311,312,313}: pytest-asyncio
     requests
     py38: pygelf
 setenv =
-    py{38,39,310,311,312}: COVERAGE_FILE=.coverage.{envname}
+    py{38,39,310,311,312,313}: COVERAGE_FILE=.coverage.{envname}
 commands =
     pytest --cov=odin_control --cov-report=term-missing --asyncio-mode=strict {posargs:-vv}
 depends =
-    py{38,39,310,311,312}: clean
-    report: py{38,39,310,311,312}
+    py{38,39,310,311,312,313}: clean
+    report: py{38,39,310,311,312,313}
 
 [testenv:clean]
 skip_install = true
