+++ tests/adapters/test_async_adapter.py
@@ -1,12 +1,7 @@
-import sys
-
 import pytest
 
-if sys.version_info[0] < 3:
-    pytest.skip("Skipping async tests", allow_module_level=True)
-else:
-    from odin_control.adapters.async_adapter import AsyncApiAdapter
-    from unittest.mock import Mock
+from odin_control.adapters.async_adapter import AsyncApiAdapter
+from unittest.mock import Mock
 
 class AsyncApiAdapterTestFixture(object):
