+++ tests/adapters/test_proxy.py
@@ -8,7 +8,6 @@ import builtins
 import threading
 import logging
 import time
-from io import StringIO
 
 import pytest
 
@@ -16,7 +15,6 @@ import requests
 
 from tornado.testing import bind_unused_port
 from tornado.ioloop import IOLoop
-from tornado.httpclient import HTTPResponse
 from tornado.web import Application, RequestHandler
 from tornado.httpserver import HTTPServer
 import tornado.gen
@@ -24,14 +22,10 @@ import tornado.gen
 from odin_control.adapters.proxy import ProxyTarget, ProxyAdapter
 from odin_control.adapters.parameter_tree import ParameterTree, ParameterTreeError
 from odin_control.adapters.adapter import wants_metadata
-from odin_control.util import convert_unicode_to_string
 from tests.utils import log_message_seen
 
-if sys.version_info[0] == 3:  # pragma: no cover
-    from unittest.mock import Mock, patch
-    import asyncio
-else:                         # pragma: no cover
-    from mock import Mock, patch
+from unittest.mock import Mock, patch
+import asyncio
 
 
 class ProxyTestHandler(RequestHandler):
@@ -71,7 +65,7 @@ class ProxyTestHandler(RequestHandler):
 
     def put(self, path):
         """Handle PUT requests to the test server."""
-        response_body = convert_unicode_to_string(tornado.escape.json_decode(self.request.body))
+        response_body = tornado.escape.json_decode(self.request.body)
         try:
             self.param_tree.set(path, response_body)
             data_ref = self.param_tree.get(path)
@@ -98,8 +92,7 @@ class ProxyTestServer(object):
 
     def _run_server(self):
 
-        if sys.version_info[0] == 3:
-            asyncio.set_event_loop(asyncio.new_event_loop())
+        asyncio.set_event_loop(asyncio.new_event_loop())
 
         self.server_event_loop = IOLoop()
 
@@ -383,7 +376,7 @@ class TestProxyAdapter():
         for tgt in range(proxy_adapter_test.num_targets):
             node_str = 'node_{}'.format(tgt)
             assert node_str in response.data
-            assert convert_unicode_to_string(response.data[node_str]) == ProxyTestHandler.param_tree.get("")
+            assert response.data[node_str] == ProxyTestHandler.param_tree.get("")
 
     def test_adapter_get_proxy_path(self, proxy_adapter_test):
         """Test that a GET to a sub-path within a targer succeeds and return the correct data."""
@@ -421,7 +414,7 @@ class TestProxyAdapter():
 
         logging.debug("Response: %s", response.data)
         assert proxy_adapter_test.adapter.param_tree.get('')['status'][node]['status_code'] == 200
-        assert convert_unicode_to_string(response.data["more"]["replace"]) == "been replaced"
+        assert response.data["more"]["replace"] == "been replaced"
 
     def test_adapter_get_bad_path(self, proxy_adapter_test):
         """Test that a GET to a bad path within a target returns the appropriate error."""
